requirements.txt
python-dateutil
slack_sdk
python-dotenv
